Fix bugs, partially (see 8bae7c0).
The problem/solution anticipated in this previous commit was not
the actual issue (rather caused by my misunderstanding of a
particular aspect of Racket modules). These changes, however, do
solve (or circumvent) it. (That said, there are still some problems
left in the implementation of CU, at least with bound variables.)