Require explicit marking of jump positions.
This works, and the marking construct is like
callcc. The "jump" fragment now captures the
essential aspects of the implicit target language
from "Refunctionalization at Work" (Danvy and
Millikan 2007)'s Dyck word recognizer example.
I.e., it does not have first-class functions, but
callcc as a (for the purpose of refunctionalizing
the recognizer) sufficient approximation.